+++ CHANGELOG.md
@@ -78,6 +78,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#580](https://github.com/babylonlabs-io/finality-provider/pull/580) chore: bump bbn for v2 compatibility
 * [#572](https://github.com/babylonlabs-io/finality-provider/pull/572)
   chore(rollup): ensure pub randomness is timestamped
+* [#583](https://github.com/babylonlabs-io/finality-provider/pull/583)
+  feat(rollup): sparse pub rand generation
 * [#584](https://github.com/babylonlabs-io/finality-provider/pull/584) chore: bump go 1.24
 * [#586](https://github.com/babylonlabs-io/finality-provider/pull/586) chore: init fp metrics